Add files via upload
Combining User-Interface elements through Python Pandas and Matplotlib, with a CSV back-end, this project aims to wrangle, re-structure, render, analyze and interpret real-time world-wide vaccination data imported from Kaggle. It enables user-interactivity by providing the functionality to append user data in specified feeds to the existing dataset, without compromising data integrity and consistency. The project is a system that can help users, analysts and researchers to understand the status of vaccination progress over the period of a year and gain significant insights.